cleanup more tests and spaces
