+++ node/network/bridge/src/tests.rs
@@ -1252,11 +1252,11 @@ fn spread_event_to_subsystems_is_up_to_date() {
 			AllMessages::GossipSupport(_) => unreachable!("Not interested in network events"),
 			AllMessages::DisputeCoordinator(_) => unreachable!("Not interested in network events"),
 			AllMessages::DisputeParticipation(_) => unreachable!("Not interetsed in network events"),
-            // Add variants here as needed, `{ cnt += 1; }` for those that need to be
-            // notified, `unreachable!()` for those that should not.
-        }
-    }
-    assert_eq!(cnt, EXPECTED_COUNT);
+			// Add variants here as needed, `{ cnt += 1; }` for those that need to be
+			// notified, `unreachable!()` for those that should not.
+		}
+	}
+	assert_eq!(cnt, EXPECTED_COUNT);
 }
 
 #[test]
